requirements.txt
langchain>=0.1.0
langchain_community>=0.0.13
langchain_core>=0.1.10
langchain_openai>=0.0.5
langgraph>=0.0.20
python-dotenv>=1.0.0
tiktoken>=0.5.2
pydantic>=2.5.2
numpy>=1.26.0
matplotlib>=3.8.0
python-dateutil>=2.8.2
SQLAlchemy>=2.0.23